Added confirmation dialog before topic deletion

+++ app/assets/javascripts/topics.coffee
@@ -4,14 +4,16 @@
 
 $(document).on 'load page:ready page:change', ->
   $('.destroy-action').click ->
-    topic_delte = $(this).closest('li')
-    delete_from_left_menu = $("##{ $(this).prev('.delete-link').attr('href').substring(8) }")
-    $.ajax({
-      url: $(this).prev('.delete-link').attr 'href'
-      type: 'DELETE'
-      dataType: 'json'
-      success: ->
-        topic_delte.remove()
-        #removes the topic from the topic menu
-        delete_from_left_menu.remove()
-    })
+    confirmation = confirm('Are you sure?')
+    if confirmation
+      topic_delte = $(this).closest('li')
+      delete_from_left_menu = $("##{ $(this).prev('.delete-link').attr('href').substring(8) }")
+      $.ajax({
+        url: $(this).prev('.delete-link').attr 'href'
+        type: 'DELETE'
+        dataType: 'json'
+        success: ->
+          topic_delte.remove()
+          #removes the topic from the topic menu
+          delete_from_left_menu.remove()
+      })
